Undo prettier workaround
