+++ packages/chat-stateful-client/src/index.ts
@@ -5,5 +5,12 @@ export { createStatefulChatClient } from './StatefulChatClient';
 
 export type { StatefulChatClient, StatefulChatClientArgs, StatefulChatClientOptions } from './StatefulChatClient';
 export type { ChatMessageWithStatus } from './types/ChatMessageWithStatus';
-export type { ChatClientState, ChatErrors, ChatThreadClientState, ChatThreadProperties } from './ChatClientState';
-export type { ErrorTargets, MethodName, ObjectMethodNames } from './ObjectMethodNames';
+export type {
+  ChatClientState,
+  ChatErrors,
+  ChatThreadClientState,
+  ChatThreadProperties,
+  ErrorTargets,
+  MethodName,
+  ObjectMethodNames
+} from './ChatClientState';
